# contract/openapi.json
{
  "components": {
    "schemas": {
      "V1Envelope": {
        "description": "The v1 success envelope: apiVersion + requestId (echoed as X-Request-Id) around the data payload.",
        "properties": {
          "apiVersion": {
            "const": "v1",
            "type": "string"
          },
          "requestId": {
            "type": "string"
          }
        },
        "required": [
          "apiVersion",
          "requestId"
        ],
        "type": "object"
      },
      "V1ErrorEnvelope": {
        "description": "The v1 error envelope: a stable code, a human message, and optional params.",
        "properties": {
          "apiVersion": {
            "const": "v1",
            "type": "string"
          },
          "error": {
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "params": {
                "type": "object"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "type": "object"
          },
          "requestId": {
            "type": "string"
          }
        },
        "required": [
          "apiVersion",
          "requestId",
          "error"
        ],
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Generated from the side-effect-free route manifest in internal/contract. Every response wraps in the v1 envelope; the envelope is documented once in components and referenced everywhere.",
    "title": "Janusly — v1 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/run": {
      "get": {
        "operationId": "get_v1_run",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/V1Envelope"
                    },
                    {
                      "properties": {
                        "data": {
                          "properties": {
                            "events": {
                              "items": {
                                "type": "object"
                              },
                              "type": "array"
                            },
                            "eventsCursor": {
                              "type": [
                                "string",
                                "null"
                              ]
                            },
                            "eventsHasMore": {
                              "type": "boolean"
                            },
                            "nodes": {
                              "items": {
                                "type": "object"
                              },
                              "type": "array"
                            },
                            "run": {
                              "type": "object"
                            }
                          },
                          "type": "object"
                        }
                      },
                      "type": "object"
                    }
                  ]
                }
              }
            },
            "description": "Success — the data payload inside the v1 envelope"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrorEnvelope"
                }
              }
            },
            "description": "Error — the v1 error envelope"
          }
        },
        "summary": "One run with nodes and paginated events"
      }
    },
...
    "/runs": {
      "get": {
        "operationId": "get_v1_runs",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/V1Envelope"
                    },
                    {
                      "properties": {
                        "data": {
                          "items": {
                            "type": "object"
                          },
                          "type": "array"
                        }
                      },
                      "type": "object"
                    }
                  ]
                }
              }
            },
            "description": "Success — the data payload inside the v1 envelope"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrorEnvelope"
                }
              }
            },
            "description": "Error — the v1 error envelope"
          }
        },
        "summary": "Keyset-paginated run list"
      }
    },
...
    "/status": {
      "get": {
        "operationId": "get_v1_status",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/V1Envelope"
                    },
                    {
                      "properties": {
                        "data": {
                          "properties": {
                            "events": {
                              "items": {
                                "type": "object"
                              },
                              "type": "array"
                            },
                            "eventsCursor": {
                              "type": [
                                "string",
                                "null"
                              ]
                            },
                            "eventsHasMore": {
                              "type": "boolean"
                            },
                            "nodes": {
                              "items": {
                                "type": "object"
                              },
                              "type": "array"
                            },
                            "run": {
                              "type": "object"
                            }
                          },
                          "type": "object"
                        }
                      },
                      "type": "object"
                    }
                  ]
                }
              }
            },
            "description": "Success — the data payload inside the v1 envelope"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrorEnvelope"
                }
              }
            },
            "description": "Error — the v1 error envelope"
          }
        },
        "summary": "Alias of /v1/run"
      }
    },
...
  },
  "servers": [
    {
      "description": "Version 1",
      "url": "/v1"
    }
  ]
}